Tests reorganization (Python)
